third_party/blink/web_tests/external/wpt/dom/nodes/moveBefore/moveBefore-size-query-expected.txt

This is a testharness.js-based test.
[FAIL] moveBefore() between different size containers invalidates target
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
